test(withTableSettings): update tests
